Port TestChronoLocalDate to scalatest
