+++ js/src/test/scala/org/threeten/bp/Platform.scala
@@ -3,6 +3,7 @@ package org.threeten.bp
 object Platform {
   type NPE = scala.scalajs.js.JavaScriptException
   type DFE = scala.scalajs.runtime.UndefinedBehaviorError
+  type CCE = scala.scalajs.runtime.UndefinedBehaviorError
 
   /** Returns `true` if and only if the code is executing on a JVM.
    *  Note: Returns `false` when executing on any JS VM.
